+++ src/process.ts
@@ -11,20 +11,25 @@ import sizeOf from 'image-size';
 import { autoUpdater } from 'electron-updater';
 import { format, formatDistanceToNowStrict } from 'date-fns';
 import { ChildProcess, fork } from 'child_process';
-import { Channel, Mode, ProcessType } from 'kit-bridge/cjs/enum';
-import { Choice, MessageData, Script, PromptData } from 'kit-bridge/cjs/type';
-import { getAppDb } from 'kit-bridge/cjs/db';
+import { Channel, Mode, ProcessType } from '@johnlindquist/kit/cjs/enum';
+import {
+  Choice,
+  MessageData,
+  Script,
+  PromptData,
+} from '@johnlindquist/kit/cjs/type';
+import { getAppDb } from '@johnlindquist/kit/cjs/db';
 
 import {
   resolveToScriptPath,
   KIT_MAC_APP,
   KIT_MAC_APP_PROMPT,
-  PATH,
   kitPath,
   kenvPath,
   kitDotEnv,
   execPath,
-} from 'kit-bridge/cjs/util';
+  PROCESS_PATH,
+} from '@johnlindquist/kit/cjs/util';
 
 import { getLog } from './logs';
 import {
@@ -426,9 +431,10 @@ const createChild = ({
     execPath,
     env: {
       ...process.env,
+      NODE_NO_WARNINGS: '1',
       KIT_CONTEXT: 'app',
       KIT_MAIN: scriptPath,
-      PATH,
+      PATH: PROCESS_PATH,
       KENV: kenvPath(),
       KIT: kitPath(),
       KIT_DOTENV: kitDotEnv(),