+++ src/shortcuts.ts
@@ -1,12 +1,12 @@
 import { globalShortcut } from 'electron';
 import log from 'electron-log';
 import { readFile } from 'fs/promises';
-import { Script } from 'kit-bridge/cjs/type';
+import { Script } from '@johnlindquist/kit/cjs/type';
 import {
   mainScriptPath,
   shortcutsPath,
   shortcutNormalizer,
-} from 'kit-bridge/cjs/util';
+} from '@johnlindquist/kit/cjs/util';
 import { runPromptProcess } from './kit';
 import { emitter, KitEvent } from './events';